Create Database Schema
Based on the post request the future API endpoint will consume,
this commit creates Customer, Subscription, and Gift models. Each model's
fields also conform to the sample post request.  From the example, it
appears that a Customer can have one Subscription, but many Gift
subscriptions.  As a result, I have set up foreign key associations on
the Subscription and Gift models.

In Ruby, it is possible to have a `belongs_to` association, where I
could specify that a Customer belongs to a Subscription while a
Subscription has many Customers, but I may not need that level of
specificity here. I will revisit if my assumption is incorrect.