Update Customer Model And Serializer
While making the `create` action for the Customer serializer, I realized
that there were a few fields that needed to be updated. This commit
allows `address_2` to be blank and also adds a validation on a Gift's
`recipient_email`.

The bulk of the work here, however, was on the `create` action.
Unfortunately, DRF has a long history of complexity when it comes to
nested serializers.  Given the time constraints, I was unable to resolve
that issue, but I believe the next steps would be to try to customize
the `ListSerializer` behavior.