Replace `.add` with `.set`
The `.set` method was recommended after an error was raised using
`.add`.  It seems that when assigning a related object, this is
preferred.  Since `.set` expects a list of items and not a dictionary, I
removed the `**` before `gift_data` and `subscription_data`.  Finally, I
updated the assignment for `subscription_data` to  `validated_data.pop('subscription')`.